backend/chatbotbackend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
requests==2.31.0
google-generativeai==0.5.0
